+++ Source/Bogus.Tests/NameTests.cs
@@ -121,5 +121,11 @@ namespace Bogus.Tests
             //switch to EN
             n["en"].LastName().Should().Be("Schultz");
         }
+
+        [Test]
+        public void get_a_random_occupation()
+        {
+            name.Occupation().Should().Be("Karate Master");
+        }
     }
 }