+++ Source/Bogus/DataSets/Name.cs
@@ -166,5 +166,13 @@
         {
             return GetRandomArrayItem("title.job");
         }
+
+        /// <summary>
+        /// Get a random occupation.
+        /// </summary>
+        public string Occupation()
+        {
+            return GetRandomArrayItem("occupation");
+        }
     }
 }